Reorganize Lambda code
Still not ideal, but it's not all in handler.py at least
